% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_table_codecheckers.R
\name{add_all_codecheckers_hyperlink}
\alias{add_all_codecheckers_hyperlink}
\title{Add Hyperlinks to Codecheckers Table}
\usage{
add_all_codecheckers_hyperlink(table)
}
\arguments{
\item{table}{The codecheckers table}
}
\value{
The data frame with added hyperlinks in the specified columns.
}
\description{
Adds hyperlinks to the codecheckers table by modifying the codechecker names, 
number of codechecks, and ORCID IDs into clickable links.
}
