% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_table_venues.R
\name{add_all_venues_hyperlinks_non_reg}
\alias{add_all_venues_hyperlinks_non_reg}
\title{Add Hyperlinks to All Venues Table}
\usage{
add_all_venues_hyperlinks_non_reg(table)
}
\arguments{
\item{table}{The data frame containing data on all venues.}
}
\value{
The data frame with hyperlinks added to the appropriate columns.
}
\description{
Adds hyperlinks to the venue names, venue types, and the number of codechecks 
in the all venues table. The links point to the venue's page and the venue type's page.
}
