% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_preprocess_register.R
\name{add_check_time}
\alias{add_check_time}
\title{Function for adding check time to each report in the register table.}
\usage{
add_check_time(register_table, register)
}
\arguments{
\item{register_table}{The register table}

\item{register}{The register from the register.csv file}
}
\value{
The adjusted register table
}
\description{
Function for adding check time to each report in the register table.
}
