% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_preprocess_register.R
\name{add_codechecker}
\alias{add_codechecker}
\title{Function for adding codechecker to each report in the register table.}
\usage{
add_codechecker(register_table, register)
}
\arguments{
\item{register_table}{The register table}

\item{register}{The register from the register.csv file}
}
\value{
The adjusted register table
}
\description{
Function for adding codechecker to each report in the register table.
}
