% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_preprocess_register.R
\name{add_issue_number_links}
\alias{add_issue_number_links}
\title{Function for adding clickable links to the issue number of each report in the register table.}
\usage{
add_issue_number_links(register_table, register)
}
\arguments{
\item{register_table}{The register table}

\item{register}{The register from the register.csv file}
}
\value{
The adjusted register table
}
\description{
Function for adding clickable links to the issue number of each report in the register table.
}
