% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_register_mds.R
\name{add_markdown_title}
\alias{add_markdown_title}
\title{Function to add the markdown title based on the specific register table name.}
\usage{
add_markdown_title(filter, md_table, register_table_name)
}
\arguments{
\item{register_table_name}{The name of the register table}

\item{markdown_table}{The markdown template where the title needs to be added}
}
\value{
The modified markdown table
}
\description{
Function to add the markdown title based on the specific register table name.
}
