% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_cert_md.R
\name{add_repository_hyperlink}
\alias{add_repository_hyperlink}
\title{Replaces a placeholder in Markdown content with a hyperlink to the repository, based on the repository type (e.g., GitHub, OSF, GitLab).}
\usage{
add_repository_hyperlink(md_content, repo_link)
}
\arguments{
\item{md_content}{A character string containing Markdown content with a placeholder for the repository link.}

\item{repo_link}{A character string containing the repository URL to be converted into a hyperlink.}
}
\value{
The modified markdown content.
}
\description{
Replaces a placeholder in Markdown content with a hyperlink to the repository, based on the repository type (e.g., GitHub, OSF, GitLab).
}
