% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_create_filtered_register_csvs.R
\name{add_repository_links_csv}
\alias{add_repository_links_csv}
\title{Function for adding repository links in the register table for CSV files.}
\usage{
add_repository_links_csv(register_table)
}
\arguments{
\item{register_table}{The register table}
}
\value{
Register table with Repository Link column added
}
\description{
Function for adding repository links in the register table for CSV files.
}
