% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_reigster_html.r
\name{add_repository_links_html}
\alias{add_repository_links_html}
\title{Function for adding repository links in the register table for the creation of the html file.}
\usage{
add_repository_links_html(register_table)
}
\arguments{
\item{register_table}{The register table}
}
\value{
Register table with adjusted repository links
}
\description{
Function for adding repository links in the register table for the creation of the html file.
}
