% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_register.R
\name{add_repository_links_md}
\alias{add_repository_links_md}
\title{Function for adding repository links in the register table for the creation of the markdown file.}
\usage{
add_repository_links_md(register_table)
}
\arguments{
\item{register_table}{The register table}
}
\value{
Register table with adjusted repository links
}
\description{
Function for adding repository links in the register table for the creation of the markdown file.
}
