% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_table_venues.R
\name{add_venue_type_hyperlink}
\alias{add_venue_type_hyperlink}
\title{Add Hyperlinks to Venue Type-Specific Table}
\usage{
add_venue_type_hyperlink(table, venue_type)
}
\arguments{
\item{table}{The data frame containing the venue type-specific data.}

\item{venue_type}{A string specifying the venue type.}
}
\value{
The data frame with hyperlinks added to the appropriate columns.
}
\description{
Adds hyperlinks to the venue names and the number of codechecks 
in the venue type-specific table. The links point to the venue's page for each venue type.
}
