% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_register_general.r
\name{add_venue_type_hyperlinks_reg}
\alias{add_venue_type_hyperlinks_reg}
\title{Function for adding clickable links to the codecheck venue type pages for each entry in the register table.}
\usage{
add_venue_type_hyperlinks_reg(register_table)
}
\arguments{
\item{register_table}{The register table}
}
\value{
The adjusted register table
}
\description{
Function for adding clickable links to the codecheck venue type pages for each entry in the register table.
}
