% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecheck.R
\name{as_latex_url}
\alias{as_latex_url}
\title{Wrap URL for LaTeX}
\usage{
as_latex_url(x)
}
\arguments{
\item{url}{- TRUE to keep relative pathname of figures.}
}
\value{
A string with the passed url as a latex `\url{}`
}
\description{
Wrap URL for LaTeX
}
\author{
Stephen Eglen
}
