% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_download_certs.R
\name{cert_link_cache}
\alias{cert_link_cache}
\title{In-session cache of resolved certificate download links, see get_cert_link()}
\usage{
cert_link_cache
}
\description{
In-session cache of resolved certificate download links, see get_cert_link()
}
