% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_register_check.R
\name{check_certificate_id}
\alias{check_certificate_id}
\title{Function for checking ceritificate id if there is a codecheck_yaml.
If there is a mismatch a stop is sent. Else a warning is thrown.}
\usage{
check_certificate_id(entry, codecheck_yaml)
}
\arguments{
\item{entry}{The registry entry}

\item{codecheck_yaml}{The codecheck yaml file}
}
\value{
None
}
\description{
Function for checking ceritificate id if there is a codecheck_yaml.
If there is a mismatch a stop is sent. Else a warning is thrown.
}
