% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_register_check.R
\name{check_issue_status}
\alias{check_issue_status}
\title{Function issue status. If the issue is not closed a warning is thrown
stating that the issue is still open.}
\usage{
check_issue_status(entry)
}
\arguments{
\item{entry}{The codecheck entry}
}
\value{
None
}
\description{
Function issue status. If the issue is not closed a warning is thrown
stating that the issue is still open.
}
