% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_register_check.R
\name{check_repository_license}
\alias{check_repository_license}
\title{Check that a repository has a license}
\usage{
check_repository_license(entry, spec)
}
\arguments{
\item{entry}{The registry entry}

\item{spec}{The parsed repository spec, see `parse_repository_spec()`}
}
\value{
None
}
\description{
Not required by the CODECHECK spec, but good practice, see the TODO in
`register_check()`. Informational only.
}
