% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecheck.R
\name{citation}
\alias{citation}
\title{Print a citation for the codecheck certificate.}
\usage{
citation(metadata)
}
\arguments{
\item{metadata}{- the codecheck metadata list.}
}
\value{

}
\description{
Print a citation for the codecheck certificate.
}
\details{
Turn the metadata into a readable citation for this document.
}
\author{
Stephen Eglen
}
