% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecheck.R
\name{cite_certificate}
\alias{cite_certificate}
\title{Print a citation for the codecheck certificate.}
\usage{
cite_certificate(metadata)
}
\arguments{
\item{metadata}{- the codecheck metadata list.}
}
\description{
Print a citation for the codecheck certificate.
}
\details{
Turn the metadata into a readable citation for this document.
}
\author{
Stephen Eglen
}
