% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_download_certs.R
\name{clear_cert_link_cache}
\alias{clear_cert_link_cache}
\title{Clear the in-session cache of resolved certificate download links}
\usage{
clear_cert_link_cache()
}
\description{
Clear the in-session cache of resolved certificate download links
}
