% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{package}
\name{codecheck-package}
\alias{codecheck}
\alias{codecheck-package}
\title{codecheck: Helper Functions for CODECHECK Project}
\description{
This contains helper functions for codecheckers for conducting CODECHECKs, e.g., authoring CODECHECK certificates (\url{https://codecheck.org.uk}), and the rendering of the CODECHECK Register (\url{https://codecheck.org.uk/register/}).
}
\seealso{
Useful links:
\itemize{
  \item \url{http://codecheck.org.uk/codecheck/}
  \item \url{https://github.com/codecheckers/codecheck}
}

}
\author{
\strong{Maintainer}: Stephen Eglen \email{sje30@cam.ac.uk} (\href{https://orcid.org/0000-0001-8607-8025}{ORCID})

Authors:
\itemize{
  \item Stephen Eglen \email{sje30@cam.ac.uk} (\href{https://orcid.org/0000-0001-8607-8025}{ORCID})
  \item Daniel Nuest \email{daniel.nuest@tu-dresden.de} (\href{https://orcid.org/0000-0002-0024-5046}{ORCID})
  \item Angelina Momin \email{angelina.momin53@gmail.com} (\href{https://orcid.org/0000-0002-6875-1551}{ORCID})
  \item Sam Langton \email{s.h.langton@amsterdamumc.nl} (\href{https://orcid.org/0000-0002-1322-1553}{ORCID})
}

}
\keyword{internal}
