% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecheck.R
\name{codecheck_metadata}
\alias{codecheck_metadata}
\title{Return the metadata for the codecheck project in root folder of project}
\usage{
codecheck_metadata(root = getwd())
}
\arguments{
\item{root}{Path to the root folder of the project, defaults to current working directory}
}
\value{
A list containing the metadata found in the codecheck.yml file
}
\description{
Return the metadata for the codecheck project in root folder of project
}
\author{
Stephen Eglen
}
