% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_codecheckers.R
\name{CODECHECKER_LIST_CACHE_DIRS}
\alias{CODECHECKER_LIST_CACHE_DIRS}
\title{Cache subdirectory of the last codechecker lists fetched successfully}
\usage{
CODECHECKER_LIST_CACHE_DIRS
}
\description{
Cache subdirectory of the last codechecker lists fetched successfully
}
\keyword{internal}
