% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_codecheckers.R
\name{CODECHECKER_LIST_COLUMNS}
\alias{CODECHECKER_LIST_COLUMNS}
\title{The columns a codechecker list is normalised to}
\usage{
CODECHECKER_LIST_COLUMNS
}
\description{
The columns a codechecker list is normalised to
}
\keyword{internal}
