% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_organisations.R
\name{commons_file_url}
\alias{commons_file_url}
\title{The URL a Wikimedia Commons file is served from}
\usage{
commons_file_url(file, width = 320)
}
\arguments{
\item{file}{A Commons file name, e.g. "Zegel Technische Universiteit Delft.svg"}

\item{width}{The width to render the file at, in pixels}
}
\value{
The Special:FilePath URL
}
\description{
The URL a Wikimedia Commons file is served from
}
\keyword{internal}
