% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_table_venues.R
\name{create_all_venues_table}
\alias{create_all_venues_table}
\title{Create All Venues Table}
\usage{
create_all_venues_table(register_table)
}
\arguments{
\item{register_table}{The data frame containing the original register data.}
}
\value{
A data frame with venues, their types, the number of codechecks, and a slug name.
}
\description{
This function generates a table with all unique venues and their corresponding types.
It also adds columns for the number of codechecks and a slug for the venue name.
}
