% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_create_filtered_register_csvs.R
\name{create_codechecker_filtered_reg_csv}
\alias{create_codechecker_filtered_reg_csv}
\title{Creates filtered codecheckers register CSV file}
\usage{
create_codechecker_filtered_reg_csv()
}
\description{
This function reads a temporary codechecker register, filters the data based on 
ORCID IDs, and saves the filtered registers as CSV files. Each CSV file is saved 
in a directory named after the corresponding ORCID ID.
}
