% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_filter_register_table.R
\name{create_codecheckers_filtered_register_tables}
\alias{create_codecheckers_filtered_register_tables}
\title{Create filtered codechecker register table}
\usage{
create_codecheckers_filtered_register_tables(register_table)
}
\arguments{
\item{register_table}{The original unfiltered register table.}
}
\value{
A list where each key corresponds to a codechecker (identified by ORCID ID) and the value is 
        a DataFrame containing the filtered data for that codechecker.
}
\description{
Filters a register based on ORCID ID and creates a register table for each ORCID ID
Each filtered table is stored in a list structure where the keys are the 
ORCID
}
