% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_create_filtered_register_csvs.R
\name{create_filtered_reg_csvs}
\alias{create_filtered_reg_csvs}
\title{Creates filtered CSV files from a register based on specified filters.}
\usage{
create_filtered_reg_csvs(register, filter_by)
}
\arguments{
\item{register}{The register to be filtered.}

\item{filter_by}{List of filters to apply (e.g., "venues", "codecheckers").}
}
\description{
The function processes the register by applying filters specified in `filter_by`. 
For "codecheckers", a temporary CSV is loaded and processed as the original register.csv
does not have the codechecker column. 
The register is then grouped by the filter column, and for each group, a CSV file is generated.
}
