% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_create_filtered_register_csvs.R
\name{create_filtered_register_csvs}
\alias{create_filtered_register_csvs}
\title{Creates filtered register csv files}
\usage{
create_filtered_register_csvs(filter_by, register)
}
\arguments{
\item{filter_by}{A vector of strings specifying the names of the columns to filter by.}

\item{register}{A dataframe representing the register data to be filtered.}
}
\description{
Each csv file is saved in the appropriate output_dir.
}
