% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_filter_register_table.R
\name{create_filtered_register_tables}
\alias{create_filtered_register_tables}
\title{Function for creating a list of register tables that is filtered based on the filter type.
Each entry in the resulting list is the filtered register table with index corresponding to unique values in that column.}
\usage{
create_filtered_register_tables(register_table, filter_by)
}
\arguments{
\item{register_table}{The register table}

\item{filter}{The filter name}
}
\value{
A list of filtered register tables
}
\description{
Function for creating a list of register tables that is filtered based on the filter type.
Each entry in the resulting list is the filtered register table with index corresponding to unique values in that column.
}
