% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_register_htmls.R
\name{create_index_postfix_html}
\alias{create_index_postfix_html}
\title{Dynamically generates the index_postfix.html from a template file}
\usage{
create_index_postfix_html(output_dir, filter, table_details)
}
\arguments{
\item{output_dir}{The output directory}

\item{filter}{The filter name}

\item{table_details}{List containing details such as the table name, subcat name.}
}
\description{
Dynamically generates the index_postfix.html from a template file
}
