% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_external_libs.R
\name{create_libs_readme}
\alias{create_libs_readme}
\title{Create README for Libraries Directory}
\usage{
create_libs_readme(libs_dir, provenance)
}
\arguments{
\item{libs_dir}{Libraries directory}

\item{provenance}{Provenance data frame}
}
\description{
Create README for Libraries Directory
}
\keyword{internal}
