% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_register_sort.R
\name{create_list_venue_sorted_register_tables}
\alias{create_list_venue_sorted_register_tables}
\title{Function for creating a list of register tables that is sorted by venue. 
The resulting list is of the form ["venue_{venue_name}": sorted_register_table]}
\usage{
create_list_venue_sorted_register_tables(register_table)
}
\arguments{
\item{register_table}{The register table}
}
\value{
A list of venue sorted register tables.
}
\description{
Function for creating a list of register tables that is sorted by venue. 
The resulting list is of the form ["venue_{venue_name}": sorted_register_table]
}
