% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_table_non_registers.R
\name{create_non_register_files}
\alias{create_non_register_files}
\title{Create Non-Register Files}
\usage{
create_non_register_files(register_table, filter_by)
}
\arguments{
\item{register_table}{The original register data.}

\item{filter_by}{A list specifying the filters to apply (e.g., "venues", "codecheckers").}
}
\description{
Processes the register table to create non-register files such as venues and codecheckers.
It applies filters to split the data into separate tables and generates corresponding HTML and JSON files.
}
