% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_table_non_registers.R
\name{create_tables_non_register}
\alias{create_tables_non_register}
\title{Create Non-Register Tables}
\usage{
create_tables_non_register(register_table, filter)
}
\arguments{
\item{register_table}{The original register data.}

\item{filter}{A string specifying the filter to apply (e.g., "venues", "codecheckers").}
}
\value{
A list of tables generated based on the specified filter. The keys are the table
names and the values are the tables themselves.
}
\description{
Generates tables based on the filter type provided, such as venues or codecheckers.
It creates tables for further processing and rendering into different formats.
}
