% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_preprocess_register.R
\name{create_temp_register_with_codechecker}
\alias{create_temp_register_with_codechecker}
\title{Creates a temporary CSV register with a "Codechecker" column.}
\usage{
create_temp_register_with_codechecker(register_table)
}
\arguments{
\item{register_table}{The register table with a "Codechecker" column.}
}
\description{
The function flattens the "Codechecker" column and saves the resulting table 
as a temporary CSV file. This tempeorary CSV is needed to filter the registers b
by codecheckers.
}
