% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_create_filtered_register_csvs.R
\name{create_venue_filtered_reg_csv}
\alias{create_venue_filtered_reg_csv}
\title{Create Filtered Venue Register CSV Files}
\usage{
create_venue_filtered_reg_csv(register)
}
\arguments{
\item{register}{A DataFrame representing the complete register data to be filtered.}
}
\description{
This function filters a register DataFrame based on venue-related columns ("Type" and "Venue"). 
It saves filtered CSV files for each unique value within these columns. 
For venue names, the corresponding venue type is also passed to determine the appropriate output directory.
}
