% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_filter_register_table.R
\name{create_venues_filtered_register_tables}
\alias{create_venues_filtered_register_tables}
\title{Create filtered venue register tables}
\usage{
create_venues_filtered_register_tables(register_table)
}
\arguments{
\item{register_table}{The original unfiltered register table.}
}
\value{
A nested list where each top-level key corresponds to a venue type and each second-level key corresponds 
        to a specific venue. The value at each key is a DataFrame containing the filtered data for that venue.
}
\description{
Filters a register based on unique venue names and creates a register table for each venue name.
Each filtered table is stored in a nested list structure where the top-level keys are the 
venue types and the second-level keys are the venue names.
}
