% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_table_venues.R
\name{create_venues_tables}
\alias{create_venues_tables}
\title{Create Venues Tables}
\usage{
create_venues_tables(register_table)
}
\arguments{
\item{register_table}{The original register data.}
}
\value{
A list of tables including the general venues table and venue type-specific tables.
}
\description{
Generates tables for all venues and venues categorized by type. 
It creates a general venues table and specific venue type tables.
}
