% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikibase.R
\name{create_wikibase_entity}
\alias{create_wikibase_entity}
\title{Create one entity on the CODECHECK Wikibase}
\usage{
create_wikibase_entity(session, row, mapping_property)
}
\arguments{
\item{session}{a session from [wikibase_session()]}

\item{row}{one row of the plan from [plan_wikibase_entities()]}

\item{mapping_property}{the local id of the mapping property, or `NA` when
creating the mapping property itself}
}
\value{
the new entity's local id
}
\description{
Create one entity on the CODECHECK Wikibase
}
\keyword{internal}
