% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecheck.R
\name{create_zenodo_record}
\alias{create_zenodo_record}
\title{Create a new Zenodo record and return its pre-assigned DOI}
\usage{
create_zenodo_record(zen)
}
\arguments{
\item{zen}{- Object from zen4R to interact with Zenodo}
}
\value{
Number of zenodo record created.
}
\description{
Create a new Zenodo record and return its pre-assigned DOI
}
\details{
Run this only once per new codecheck.
}
\author{
Stephen Eglen
}
