% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_create_filtered_register_csvs.R
\name{determine_filter_column_name}
\alias{determine_filter_column_name}
\title{Determines the register table's column name to filter the data by.}
\usage{
determine_filter_column_name(filter)
}
\arguments{
\item{filter}{The filter name}
}
\value{
The column name to filter by
}
\description{
Determines the register table's column name to filter the data by.
}
