% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_register.R
\name{determine_output_file_directory}
\alias{determine_output_file_directory}
\title{Function to determine the appropriate output file directory based on the register table name.}
\usage{
determine_output_file_directory(register_table_name)
}
\arguments{
\item{register_table_name}{The name of the register table}
}
\value{
The path to the output file directory
}
\description{
Function to determine the appropriate output file directory based on the register table name.
}
