% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_create_filtered_register_csvs.R
\name{determine_venue_category}
\alias{determine_venue_category}
\title{Determines the venue category based on the venue_name}
\usage{
determine_venue_category(venue_name)
}
\arguments{
\item{venue_name}{The venue_name obtained from the "Type" column of the register}
}
\value{
The venue category. If the venue does not belong to any category the venue_name is returned
}
\description{
Determines the venue category based on the venue_name
}
