% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_codecheckers.R
\name{DONUT_CENTER}
\alias{DONUT_CENTER}
\alias{DONUT_R_OUTER}
\alias{DONUT_R_INNER}
\title{Geometry of the codechecker donut, in the 96x96 user space of its viewBox}
\usage{
DONUT_CENTER

DONUT_R_OUTER

DONUT_R_INNER
}
\description{
Geometry of the codechecker donut, in the 96x96 user space of its viewBox
}
\keyword{internal}
