% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_external_libs.R
\name{download_academicons_fonts}
\alias{download_academicons_fonts}
\title{Download Academicons Fonts}
\usage{
download_academicons_fonts(libs_dir, version)
}
\arguments{
\item{libs_dir}{Base libraries directory}

\item{version}{Academicons version}
}
\description{
Download Academicons Fonts
}
\keyword{internal}
