% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_download_certs.R
\name{download_cert_pdf}
\alias{download_cert_pdf}
\title{Downloads a certificate PDF from a report link and saves it locally. 
If the download link is a ZIP file, it extracts the PDF from 
the archive. Returns status based on success.}
\usage{
download_cert_pdf(report_link, cert_id)
}
\arguments{
\item{report_link}{URL of the report from which to download the certificate.}

\item{cert_id}{ID of the certificate, used for directory naming and logging.}
}
\value{
1 if the certificate is successfully downloaded and saved; otherwise, 0.
}
\description{
Downloads a certificate PDF from a report link and saves it locally. 
If the download link is a ZIP file, it extracts the PDF from 
the archive. Returns status based on success.
}
