% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_external_libs.R
\name{download_font_awesome_fonts}
\alias{download_font_awesome_fonts}
\title{Download Font Awesome Fonts}
\usage{
download_font_awesome_fonts(libs_dir, version)
}
\arguments{
\item{libs_dir}{Base libraries directory}

\item{version}{Font Awesome version}
}
\description{
Download Font Awesome Fonts
}
\keyword{internal}
