% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_external_libs.R
\name{download_library_fonts}
\alias{download_library_fonts}
\title{Download the Font Files of a Library}
\usage{
download_library_fonts(libs_dir, lib, force = FALSE)
}
\arguments{
\item{libs_dir}{Base libraries directory}

\item{lib}{Library specification with a `fonts` entry}

\item{force}{If TRUE, re-download fonts even if they already exist}
}
\value{
`TRUE` if at least one font file was downloaded
}
\description{
Download the Font Files of a Library
}
\keyword{internal}
