% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_reigster_html.r
\name{edit_html_lib_paths}
\alias{edit_html_lib_paths}
\title{Loads a html file and replaces the libs path in the html file to the libs folder in "docs/libs"
This is done so all html files can share one libs folder.}
\usage{
edit_html_lib_paths(html_file_path)
}
\arguments{
\item{html_file_path}{The path to the html file that needs to be edited.}
}
\description{
Loads a html file and replaces the libs path in the html file to the libs folder in "docs/libs"
This is done so all html files can share one libs folder.
}
