% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_build_metadata.R
\name{escape_html_attribute}
\alias{escape_html_attribute}
\title{Escape a value for use in an HTML attribute}
\usage{
escape_html_attribute(x)
}
\arguments{
\item{x}{A character value}
}
\value{
The value with `&`, `<`, `>`, `"` and `'` replaced by entities
}
\description{
The values of `<meta>` tags are HTML attributes, and Google Scholar's
indexing guidelines are explicit that they have to be escaped,
<https://scholar.google.com/intl/en/scholar/inclusion.html#indexing>.
Certificate metadata routinely contains ampersands and quotes, in paper
titles as much as in the free text summary of a check.
}
\keyword{internal}
