% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fediverse.R
\name{fediverse_profile_url}
\alias{fediverse_profile_url}
\title{The profile URL of a fediverse account}
\usage{
fediverse_profile_url(handle)
}
\arguments{
\item{handle}{An account, see [fediverse_handle()].}
}
\value{
The URL, or `NULL` when `handle` is not an account.
}
\description{
`https://instance/@user`, the form Mastodon and most other servers answer
to.
}
\keyword{internal}
