% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_build_metadata.R
\name{generate_footer_build_info}
\alias{generate_footer_build_info}
\title{Generate Footer Build Information HTML}
\usage{
generate_footer_build_info(metadata)
}
\arguments{
\item{metadata}{Build metadata from get_build_metadata()}
}
\value{
HTML string with build information (without wrapper styling)
}
\description{
Creates HTML content for displaying build information in the footer.
Styling should be applied in the template or CSS file.
}
