% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_register_htmls.R
\name{generate_href}
\alias{generate_href}
\title{Generate full href for for different href types.}
\usage{
generate_href(filter, table_details, href_type)
}
\arguments{
\item{filter}{The filter name}

\item{table_details}{A list containing details needed for the hrefs}

\item{href_type}{The href type (e.g., 'csv_source', 'searchable_csv', 'json', 'md')}
}
\value{
String representing the full URL to access the specified resource
}
\description{
Generate full href for for different href types.
}
