% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_table_non_registers.R
\name{generate_html_data}
\alias{generate_html_data}
\title{Generates a list of data for the html. The list contains the html
title, subtext and extra text.}
\usage{
generate_html_data(table, page_type, table_name)
}
\arguments{
\item{table}{The table to showcase in the html}

\item{page_type}{The HTML page type that needs to rendered}

\item{table_name}{The name of the table}
}
\value{
A list of the html data such as title, subtext etc
}
\description{
Generates a list of data for the html. The list contains the html
title, subtext and extra text.
}
