% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_table_non_registers.R
\name{generate_html_extra_text_non_register}
\alias{generate_html_extra_text_non_register}
\title{Generates the extra text of the HTML pages for non registers.
This extra text is to be placed under the table.
There is only extra text for the codecheckers HTML page to explain
the reason for discrepancy between total_codechecks != SUM(no.of codechecks)}
\usage{
generate_html_extra_text_non_register(page_type)
}
\arguments{
\item{page_type}{The HTML page type that needs to rendered}
}
\value{
The extra text to place under the table
}
\description{
Generates the extra text of the HTML pages for non registers.
This extra text is to be placed under the table.
There is only extra text for the codecheckers HTML page to explain
the reason for discrepancy between total_codechecks != SUM(no.of codechecks)
}
