% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_table_non_registers.R
\name{generate_html_postfix_hrefs_non_reg}
\alias{generate_html_postfix_hrefs_non_reg}
\title{Generates postfix hrefs for the venues/ codecheckers list pages}
\usage{
generate_html_postfix_hrefs_non_reg(filter, table_name)
}
\arguments{
\item{filter}{The filter being used such as "venues" or "codecheckers"}

\item{table_name}{The name of the table}
}
\value{
A list of the hrefs.
}
\description{
Generates postfix hrefs for the venues/ codecheckers list pages
}
