% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_register_htmls.R
\name{generate_html_postfix_hrefs_reg}
\alias{generate_html_postfix_hrefs_reg}
\title{Generates the hrefs to set in the postfix.html file for the rendering of register tables.}
\usage{
generate_html_postfix_hrefs_reg(
  filter,
  register_table_name,
  filter_subcategory = NULL
)
}
\arguments{
\item{filter}{The filter name}

\item{register_table_name}{The register table name}

\item{filter_subcategory}{An optional string representing a subcategory within the filter 
(e.g., venue type). Default is NULL.}
}
\value{
A list with the hrefs
}
\description{
Generates the hrefs to set in the postfix.html file for the rendering of register tables.
}
