% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_table_non_registers.R
\name{generate_html_subtext_non_register}
\alias{generate_html_subtext_non_register}
\title{Generates the subtext of the HTML pages for non registers with a summary of
the number of codechecks and number of codechecks/ venues etc.}
\usage{
generate_html_subtext_non_register(table, page_type, table_name)
}
\arguments{
\item{table}{The table to showcase in the html}

\item{page_type}{The HTML page type that needs to rendered}

\item{table_name}{The name of the table}
}
\value{
The subtext to put under the html title
}
\description{
Generates the subtext of the HTML pages for non registers with a summary of
the number of codechecks and number of codechecks/ venues etc.
}
