% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_table_non_registers.R
\name{generate_html_title_non_registers}
\alias{generate_html_title_non_registers}
\title{Generates the titles of the HTML pages for non registers}
\usage{
generate_html_title_non_registers(page_type, table_name)
}
\arguments{
\item{page_type}{The HTML page type that needs to rendered}

\item{table_name}{The name of the table}
}
\value{
The title to put on the html page
}
\description{
Generates the titles of the HTML pages for non registers
}
