% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_build_metadata.R
\name{generate_meta_generator_tag}
\alias{generate_meta_generator_tag}
\title{Generate HTML Meta Generator Tag}
\usage{
generate_meta_generator_tag(metadata)
}
\arguments{
\item{metadata}{Build metadata from get_build_metadata()}
}
\value{
HTML string with meta generator tag
}
\description{
Creates an HTML meta generator tag with build information.
}
