% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_table_non_registers.R
\name{generate_non_reg_html}
\alias{generate_non_reg_html}
\title{Generates non register html page.}
\usage{
generate_non_reg_html(table, table_name, page_type, html_data, output_dir)
}
\arguments{
\item{table}{The table to showcase in the html}

\item{table_name}{The name of the table}

\item{page_type}{The HTML page type that needs to rendered.}

\item{html_data}{A list containing the title, subtext, extra text of the html page}

\item{output_dir}{The directory where the html needs to be saved}
}
\description{
Generates non register html page.
}
