% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_register_general.r
\name{generate_output_dir}
\alias{generate_output_dir}
\title{Generate Output Directory Path}
\usage{
generate_output_dir(filter, table_details = list())
}
\arguments{
\item{filter}{The filter name (e.g., "venues", "codecheckers").}

\item{table_details}{List containing details such as the table name, subcat name.}
}
\value{
A string representing the directory path for saving files.
}
\description{
Generates the directory path for saving files based on info in the register table
and the filter name. Creates the output directory if it does not already exist
}
