% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_register_general.R
\name{generate_table_details}
\alias{generate_table_details}
\title{Generate Table Details}
\usage{
generate_table_details(table_key, table, filter, is_reg_table = TRUE)
}
\arguments{
\item{table_key}{The key (name) of the table being processed.}

\item{table}{The data frame containing the table data.}

\item{filter}{A string specifying the filter applied to the table data.}

\item{is_reg_table}{A boolean indicating whether the table is a register table (default is TRUE).}
}
\value{
A list of table details including name, slugified name, subcategory (if applicable), and output directory.
}
\description{
This function generates metadata and details for a specific table, including 
the table name, slugified name, subcategory (if applicable), and the output directory. 
It is used when rendering tables in different formats.
}
