% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_table_non_registers.R
\name{generate_table_details_non_reg}
\alias{generate_table_details_non_reg}
\title{Generate Table Details for Non-Register Files}
\usage{
generate_table_details_non_reg(table, filter, subcat = NULL)
}
\arguments{
\item{table}{The data frame containing the filtered table.}

\item{filter}{A string specifying the filter applied to the table.}

\item{subcat}{An optional string for the subcategory (if applicable).}
}
\value{
A list containing metadata such as title, subtext, and output directory for the table.
}
\description{
Generates metadata for non-register tables, including subcategory, title, subtext, 
extra text, and output directory. It is used to prepare the table for rendering.
}
