% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_cert_md.R
\name{get_abstract_text_crossref}
\alias{get_abstract_text_crossref}
\title{Extracts the paper DOI from the config_yml of the paper, 
constructs a CrossRef API request, and returns the abstract text if available.}
\usage{
get_abstract_text_crossref(register_repo)
}
\arguments{
\item{register_repo}{URL or path to the repository containing the paper's configuration.}
}
\value{
The abstract text as a string if available; otherwise, NULL.
}
\description{
Extracts the paper DOI from the config_yml of the paper, 
constructs a CrossRef API request, and returns the abstract text if available.
}
