% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_cert_md.R
\name{get_abstract_text_openalex}
\alias{get_abstract_text_openalex}
\title{Retrieves the abstract of a research paper using the OpenAlex API.}
\usage{
get_abstract_text_openalex(register_repo)
}
\arguments{
\item{register_repo}{URL or path to the repository containing the paper's configuration.}
}
\value{
The abstract text as a string if available; otherwise, NULL.
}
\description{
Retrieves the abstract of a research paper using the OpenAlex API.
}
