% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_codecheckers.R
\name{get_agile_codecheckers_data}
\alias{get_agile_codecheckers_data}
\title{Fetch and cache agile-codecheckers.csv data from GitHub}
\usage{
get_agile_codecheckers_data()
}
\value{
A data frame with columns: name, handle, ORCID, contact, fields, languages,
  fediverse (contact, fields and languages are `NA`)
}
\description{
Fetch and cache agile-codecheckers.csv data from GitHub
}
