% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_download_certs.R
\name{get_cert_link}
\alias{get_cert_link}
\title{Retrieves the download link for a certificate file from either Zenodo or OSF.}
\usage{
get_cert_link(report_link, cert_id)
}
\arguments{
\item{report_link}{URL of the report to access, either from Zenodo or OSF.}

\item{cert_id}{ID of the certificate, used for logging and warnings.}
}
\value{
The download link for the certificate file as a string if found; otherwise, NULL.
}
\description{
Retrieves the download link for a certificate file from either Zenodo or OSF.
}
