% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_download_certs.R
\name{get_cert_link_uncached}
\alias{get_cert_link_uncached}
\title{Resolves the download link for a certificate file without caching, see get_cert_link().}
\usage{
get_cert_link_uncached(report_link, cert_id)
}
\arguments{
\item{report_link}{URL of the report to access, either from Zenodo, OSF, or ResearchEquals.}

\item{cert_id}{ID of the certificate, used for logging and warnings.}
}
\value{
The download link for the certificate file as a string if found; otherwise, NULL.
}
\description{
Resolves the download link for a certificate file without caching, see get_cert_link().
}
