% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{get_codecheck_yml}
\alias{get_codecheck_yml}
\title{Get the CODECHECK configuration file from a repository}
\usage{
get_codecheck_yml(x)
}
\arguments{
\item{x}{Repository in the codecheckers organisation on GitHub}
}
\description{
Get the CODECHECK configuration file from a repository
}
\author{
Daniel Nuest
}
