% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{get_codecheck_yml_github}
\alias{get_codecheck_yml_github}
\title{Retrieve a codecheck.yml file from a GitHub repository}
\usage{
get_codecheck_yml_github(x)
}
\arguments{
\item{x}{the org/repo to download the file from}
}
\description{
Retrieve a codecheck.yml file from a GitHub repository
}
\author{
Daniel Nuest
}
