% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{get_codecheck_yml_gitlab}
\alias{get_codecheck_yml_gitlab}
\title{Retrieve a codecheck.yml file from an GitLab.com project}
\usage{
get_codecheck_yml_gitlab(x)
}
\arguments{
\item{x}{the project name on GitLab.com}
}
\description{
It seems https://statnmap.github.io/gitlabr/ always requires authentication
}
\author{
Daniel Nuest
}
