% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{get_codecheck_yml_osf}
\alias{get_codecheck_yml_osf}
\title{Retrieve a codecheck.yml file from an OSF project}
\usage{
get_codecheck_yml_osf(x)
}
\arguments{
\item{x}{the OSF id (5 characters)}
}
\description{
Retrieve a codecheck.yml file from an OSF project
}
\author{
Daniel Nüst
}
