% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{get_codecheck_yml_uncached}
\alias{get_codecheck_yml_uncached}
\title{Retrieve a codecheck.yml file from a remote repository}
\usage{
get_codecheck_yml_uncached(x)
}
\arguments{
\item{x}{the repo specification}
}
\description{
Retrieve a codecheck.yml file from a remote repository
}
\author{
Daniel Nüst
}
