% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{get_codecheck_yml_zenodo}
\alias{get_codecheck_yml_zenodo}
\title{Retrieve a codecheck.yml file from a Zenodo record}
\usage{
get_codecheck_yml_zenodo(x, sandbox = FALSE)
}
\arguments{
\item{x}{the record ID on Zenodo}

\item{sandbox}{connect with the Zenodo Sandbox instead of the real service}
}
\description{
Retrieve a codecheck.yml file from a Zenodo record
}
\author{
Daniel Nuest
}
