% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{get_github_readme_raw}
\alias{get_github_readme_raw}
\title{Retrieve the rendered README of a GitHub repository, as raw text}
\usage{
get_github_readme_raw(repo)
}
\arguments{
\item{repo}{the `org/repo` or `org/repo|path` spec}
}
\value{
the README text, or `NULL` if none was found
}
\description{
Retrieve the rendered README of a GitHub repository, as raw text
}
\author{
Daniel Nuest
}
