% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ror.R
\name{get_orcid_affiliations_cached}
\alias{get_orcid_affiliations_cached}
\title{Cached version of [get_orcid_affiliations_result()]}
\usage{
get_orcid_affiliations_cached(orcid)
}
\arguments{
\item{orcid}{An ORCID identifier (NNNN-NNNN-NNNN-NNNX).}
}
\value{
The affiliation data frame
}
\description{
Caches on disk, but only when ORCID actually answered, see
[cached_lookup()]. Cleared by [register_clear_cache()].
}
\keyword{internal}
