% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_download_certs.R
\name{get_osf_cert_link}
\alias{get_osf_cert_link}
\title{Retrieves the link to a certificate PDF file from an OSF project node. It retrieves its files, 
and searches for a single PDF certificate file within the node. If multiple or no PDF 
files are found, it returns NULL with a warning.}
\usage{
get_osf_cert_link(report_link, cert_id)
}
\arguments{
\item{report_link}{URL of the OSF report to access.}

\item{cert_id}{ID of the certificate, used for logging and warnings.}
}
\value{
The download link for the certificate file as a string if a single PDF is found; otherwise, NULL.
}
\description{
Retrieves the link to a certificate PDF file from an OSF project node. It retrieves its files, 
and searches for a single PDF certificate file within the node. If multiple or no PDF 
files are found, it returns NULL with a warning.
}
