% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cert_title.R
\name{get_osf_record_title}
\alias{get_osf_record_title}
\title{The title of an OSF project node}
\usage{
get_osf_record_title(report_link)
}
\arguments{
\item{report_link}{An OSF project URL, whose last segment is the node ID}
}
\value{
The node title, or `NULL` if the node carries none
}
\description{
The title of an OSF project node
}
\keyword{internal}
