% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_create_filtered_register_csvs.R
\name{get_output_dir}
\alias{get_output_dir}
\title{Get Output Directory Path}
\usage{
get_output_dir(filter, table_name, filter_subcategory = NULL)
}
\arguments{
\item{filter}{The filter name (e.g., "venues", "codecheckers").}

\item{table_name}{The name of the table or item being filtered.}

\item{filter_subcategory}{Optional subcategory for more specific filtering (e.g., venue type). Default is NULL.}
}
\value{
A string representing the directory path for saving files.
}
\description{
Determines the directory path for saving files based on the filter, table name, 
and an optional subcategory.
}
