% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_download_certs.R
\name{get_researchequals_cert_link}
\alias{get_researchequals_cert_link}
\title{Accesses a codecheck's ResearchEquals record via its report link and download the main file of the module}
\usage{
get_researchequals_cert_link(report_link, cert_id)
}
\arguments{
\item{report_link}{URL of the ResearchEquals report to access.}

\item{cert_id}{ID of the certificate, used for logging and warnings.}
}
\value{
The download link for the certificate file as a string if found; otherwise, NULL.
}
\description{
Accesses a codecheck's ResearchEquals record via its report link and download the main file of the module
}
