% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/researchequals.R
\name{get_researchequals_collection}
\alias{get_researchequals_collection}
\title{Retrieve a collection on ResearchEquals}
\usage{
get_researchequals_collection(
  issue_id = RESEARCHEQUALS_COLLECTIONS[[1]]$issue_id
)
}
\arguments{
\item{issue_id}{the collection issue ID, defaults to the current CODECHECK issue}
}
\value{
the parsed API response as a list, with the element `submissions`
}
\description{
Retrieve a collection from ResearchEquals
}
\details{
Fetches the issue of a collection, including the list of submissions that
constitutes its membership.
}
\author{
Daniel Nuest
}
