% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_organisations.R
\name{get_ror_record_cached}
\alias{get_ror_record_cached}
\title{Cached version of [get_ror_record_result()]}
\usage{
get_ror_record_cached(ror)
}
\arguments{
\item{ror}{A ROR id, with or without the `https://ror.org/` prefix.}
}
\value{
The parsed ROR record, or `NULL`
}
\description{
Cached version of [get_ror_record_result()]
}
\keyword{internal}
