% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_organisations.R
\name{get_wikidata_logo_cached}
\alias{get_wikidata_logo_cached}
\title{Cached version of [get_wikidata_logo_result()]}
\usage{
get_wikidata_logo_cached(qid)
}
\arguments{
\item{qid}{A Wikidata item id, e.g. "Q752663".}
}
\value{
The logo URL, or `NA_character_`
}
\description{
Cached version of [get_wikidata_logo_result()]
}
\keyword{internal}
