% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecheck.R
\name{get_zenodo_id}
\alias{get_zenodo_id}
\title{Extract the Zenodo record number from the report URL}
\usage{
get_zenodo_id(report)
}
\arguments{
\item{report}{- string containing the report URL on Zenodo.}
}
\value{
the Zenodo record number (a number with at least 7 digits).
}
\description{
Extract the Zenodo record number from the report URL
}
\details{
The report paramater should contain a URL like:
http://doi.org/10.5281/zenodo.3750741 where the final part of the
URL is zenodo.X where X is a number containing at least 7 digits.
X is returned.  If we cannot extract the number X, we return an
error, in which case the function create_zenodo_record() can be
run to create a new record.  Alternatively, the report URL is
pre-assigned a DOI when manually creating the record.
}
\author{
Stephen Eglen
}
