% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenodo.R
\name{get_zenodo_record}
\alias{get_zenodo_record}
\title{Get the full zenodo record using the record number stored in the metadata.}
\usage{
get_zenodo_record(zenodo, metadata = codecheck_metadata(getwd()))
}
\arguments{
\item{zenodo}{An object from zen4R to connect with Zenodo}

\item{metadata}{A codecheck configuration (list). Defaults to loading from
codecheck.yml in the current working directory using \code{codecheck_metadata(getwd())}.}
}
\value{
The Zenodo record, or NULL.
}
\description{
Get the full Zenodo record from the metadata
}
\details{
Retrieve the Zenodo record, if one exists. By default, loads metadata from
codecheck.yml in the current working directory.
If no record number is currently listed in the metadata (i.e. the "FIXME" tag is still there)
then the code returns NULL and an empty record should be created.
}
\author{
Stephen Eglen
}
