% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_build_metadata.R
\name{is_nonempty_string}
\alias{is_nonempty_string}
\title{Is this a usable single non-empty string?}
\usage{
is_nonempty_string(x)
}
\arguments{
\item{x}{A value from a parsed codecheck.yml, which may be NULL, NA or ""}
}
\value{
TRUE if `x` can be rendered into a meta tag
}
\description{
Is this a usable single non-empty string?
}
\keyword{internal}
