% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/researchequals.R
\name{is_researchequals_report}
\alias{is_researchequals_report}
\title{Is a report reference published on ResearchEquals?}
\usage{
is_researchequals_report(x)
}
\arguments{
\item{x}{a DOI, DOI URL or ResearchEquals URL}
}
\value{
TRUE if `x` points at ResearchEquals
}
\description{
Is a report reference published on ResearchEquals?
}
\keyword{internal}
