% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecheck.R
\name{latex_codecheck_logo}
\alias{latex_codecheck_logo}
\title{Print the latex code to include the CODECHECK logo}
\usage{
latex_codecheck_logo()
}
\description{
Print the latex code to include the CODECHECK logo
}
\author{
Stephen Eglen
}
