% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecheck.R
\name{latex_summary_of_metadata}
\alias{latex_summary_of_metadata}
\title{Print a latex table to summarise CODECHECK metadata}
\usage{
latex_summary_of_metadata(metadata)
}
\arguments{
\item{metadata}{- the codecheck metadata list.}
}
\value{
The latex table, suitable for including in the Rmd
}
\description{
Print a latex table to summarise CODECHECK metadata
}
\details{
Format a latex table that summarises the main CODECHECK metadata,
excluding the MANIFEST.
}
\author{
Stephen Eglen
}
