% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_external_libs.R
\name{library_expected_files}
\alias{library_expected_files}
\title{Expected Files of an External Library}
\usage{
library_expected_files(lib)
}
\arguments{
\item{lib}{A single library specification from [external_library_specs()]}
}
\value{
Character vector of file paths relative to the libraries directory
}
\description{
Expected Files of an External Library
}
\keyword{internal}
