% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{list_manifest_files}
\alias{list_manifest_files}
\title{Summarise manifest files.}
\usage{
list_manifest_files(root, metadata, check_dir)
}
\arguments{
\item{root}{- Path to the root folder of the project.}

\item{metadata}{- the codecheck metadata list.}

\item{check_dir}{- folder where outputs have been copied to (codecheck/outputs)}
}
\value{
A dataframe containing one row per manifest file.
}
\description{
List manifest.
}
\author{
Daniel Nuest
}
