% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_register_mds.R
\name{load_md_template}
\alias{load_md_template}
\title{Function to load the markdown table}
\usage{
load_md_template(template_path)
}
\arguments{
\item{template_path}{The path to the markdown template}
}
\value{
The markdown table template
}
\description{
Function to load the markdown table
}
