% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{parse_repository_spec}
\alias{parse_repository_spec}
\title{Parse the repository specification in the column "Repo" in the register CSV file}
\usage{
parse_repository_spec(x)
}
\arguments{
\item{x}{the repository specification to parse}
}
\value{
a named character vector with the items `type` and `repo`
}
\description{
Based roughly on [`remotes::parse_one_extra`](https://github.com/r-lib/remotes/blob/master/R/deps.R#L519)
}
\details{
Supported variants:

- `osf::ABC12`
- `github::codecheckers/Piccolo-2020`
}
\author{
Daniel Nüst
}
