% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_preprocess_register.R
\name{preprocess_register}
\alias{preprocess_register}
\title{Function for preprocessing the register to create and return the register_table.}
\usage{
preprocess_register(register)
}
\arguments{
\item{register_table}{The register table}
}
\value{
The preprocessed register table
}
\description{
Function for preprocessing the register to create and return the register_table.
}
