% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_external_libs.R
\name{register_assets_dir}
\alias{register_assets_dir}
\title{The Assets Directory Belonging to a Libraries Directory}
\usage{
register_assets_dir(libs_dir)
}
\arguments{
\item{libs_dir}{Base libraries directory}
}
\value{
the path of the assets directory
}
\description{
Both live in the render output directory, so "docs/libs" gives "docs/assets".
}
\keyword{internal}
