% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikidata_ids.R
\name{register_orcids}
\alias{register_orcids}
\title{The ORCIDs the register knows about}
\usage{
register_orcids(register_table)
}
\arguments{
\item{register_table}{the preprocessed register table}
}
\value{
the ORCIDs, unique and without `NA`
}
\description{
The ORCIDs the register knows about
}
\keyword{internal}
