% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_table_codecheckers.R
\name{render_codecheckers_json}
\alias{render_codecheckers_json}
\title{Renders JSON file of codecheckers.}
\usage{
render_codecheckers_json(list_codechecker_reg_tables)
}
\arguments{
\item{list_codechecker_reg_tables}{The list of codechecker register tables. The indices are the ORCID IDs.}
}
\description{
Renders JSON file of codecheckers.
}
