% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_table_non_registers.R
\name{render_non_register_jsons}
\alias{render_non_register_jsons}
\title{Renders JSON file of non register tables such as list of venues, list of codecheckers}
\usage{
render_non_register_jsons(list_reg_tables, page_type)
}
\arguments{
\item{list_reg_tables}{The list of register tables needed for the information.}
}
\description{
Renders JSON file of non register tables such as list of venues, list of codecheckers
}
