% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_table_non_registers.R
\name{render_non_register_md}
\alias{render_non_register_md}
\title{Render Non-Register Markdown Table}
\usage{
render_non_register_md(table, table_details, filter)
}
\arguments{
\item{table}{The data frame containing the filtered table.}

\item{table_details}{A list of metadata about the table (e.g., title, subtext, extra text).}

\item{filter}{A string specifying the filter applied (e.g., "venues", "codecheckers").}
}
\description{
Renders the table in Markdown format, including adding hyperlinks to the relevant
columns. It adjusts column widths and saves the output as a Markdown file.
}
