% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_table_non_registers.R
\name{render_non_register_tables_html}
\alias{render_non_register_tables_html}
\title{Renders non-register tables such as list of venues, codecheckers}
\usage{
render_non_register_tables_html(list_reg_tables, page_type)
}
\arguments{
\item{list_reg_tables}{The list of register tables}

\item{page_type}{The HTML page type that needs to rendered.}
}
\value{
A list of register tables. The entries in the list are the names of the table
}
\description{
Renders non-register tables such as list of venues, codecheckers
}
