% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_reigster_html.r
\name{render_register_htmls}
\alias{render_register_htmls}
\title{Renders register htmls for a list of register tables}
\usage{
render_register_htmls(list_register_tables)
}
\arguments{
\item{list_register_table}{List of register tables}
}
\description{
Renders register htmls for a list of register tables
}
