% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_register_json.R
\name{render_register_jsons}
\alias{render_register_jsons}
\title{Renders register jsons for a list of register tables}
\usage{
render_register_jsons(list_register_tables)
}
\arguments{
\item{list_register_table}{List of register tables}
}
\description{
Renders register jsons for a list of register tables
}
