% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_register.R
\name{render_register_md}
\alias{render_register_md}
\title{Render Markdown files for register tables.}
\usage{
render_register_md(list_register_tables, md_columns_widths)
}
\arguments{
\item{list_register_tables}{A list of register tables}

\item{md_columns_widths}{A string specifying the Markdown formatting for column widths.}
}
\value{
None
}
\description{
Render Markdown files for register tables.
}
