% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_register_mds.R
\name{render_register_mds}
\alias{render_register_mds}
\title{Renders register mds for a list of register tables}
\usage{
render_register_mds(list_register_tables)
}
\arguments{
\item{list_register_table}{List of register tables}
}
\description{
Renders register mds for a list of register tables
}
