% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_table_venues.R
\name{render_table_all_venues_html}
\alias{render_table_all_venues_html}
\title{Renders table for all venues in HTML format.
Each venue name links to the register table for that specific
venue. The ORCID IDs link to their ORCID pages.}
\usage{
render_table_all_venues_html(list_venue_reg_tables)
}
\arguments{
\item{list_venue_reg_tables}{The list of venue register tables. The indices are the ORCID IDs.}
}
\value{
Returns the html table of all the venues.
}
\description{
Renders table for all venues in HTML format.
Each venue name links to the register table for that specific
venue. The ORCID IDs link to their ORCID pages.
}
