% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_table_venues.R
\name{render_table_all_venues_json}
\alias{render_table_all_venues_json}
\title{Renders table for all venues in JSON format.}
\usage{
render_table_all_venues_json(list_venue_reg_tables)
}
\arguments{
\item{list_venue_reg_tables}{The list of venue register tables. The indices are the venue names.}
}
\value{
Returns a JSON table for the all venues
}
\description{
Renders table for all venues in JSON format.
}
