% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_table_codecheckers.R
\name{render_table_codecheckers_html}
\alias{render_table_codecheckers_html}
\title{Renders codecheckers table in HTML format.
Each codechecker name links to the register table for that specific
codechecker. The ORCID IDs link to their ORCID pages.}
\usage{
render_table_codecheckers_html(list_codechecker_reg_tables)
}
\arguments{
\item{list_codechecker_reg_tables}{The list of codechecker register tables. The indices are the ORCID IDs.}
}
\description{
Renders codecheckers table in HTML format.
Each codechecker name links to the register table for that specific
codechecker. The ORCID IDs link to their ORCID pages.
}
