% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_table_venues.R
\name{render_table_venues_subcat}
\alias{render_table_venues_subcat}
\title{Renders table for venue subcategories
This function can be used for both the html and json table since it works on the
table for all venues. If the all_venues table is of html format then this creates
a html table and if it is of json format, this creates a json table.}
\usage{
render_table_venues_subcat(venues_table)
}
\arguments{
\item{venues_table}{The table of all venues}
}
\value{
Returns a list of the html tables of venue subcategories. 
The names in the list are the names of the subcategories
}
\description{
Renders table for venue subcategories
This function can be used for both the html and json table since it works on the
table for all venues. If the all_venues table is of html format then this creates
a html table and if it is of json format, this creates a json table.
}
