% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_table_venues.R
\name{render_tables_venues_html}
\alias{render_tables_venues_html}
\title{Renders table for all venues and venue subcategories for HTML.}
\usage{
render_tables_venues_html(list_venue_reg_tables)
}
\arguments{
\item{list_venue_reg_tables}{The list of venue register tables. The indices are the venue names.}
}
\value{
Returns list of the html tables with the names of the list being the table names.
}
\description{
Renders table for all venues and venue subcategories for HTML.
}
