% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/researchequals.R
\name{researchequals_collection_url}
\alias{researchequals_collection_url}
\title{The public URL of a ResearchEquals collection}
\usage{
researchequals_collection_url(id)
}
\arguments{
\item{id}{the collection ID}
}
\value{
the collection URL
}
\description{
The public URL of a ResearchEquals collection
}
\keyword{internal}
