% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_organisations.R
\name{ror_external_id}
\alias{ror_external_id}
\title{One of a ROR record's external identifiers}
\usage{
ror_external_id(record, type)
}
\arguments{
\item{record}{A ROR record, or `NULL`}

\item{type}{The identifier type, e.g. "wikidata" or "grid"}
}
\value{
The preferred value, or the first one, or `NA_character_`
}
\description{
One of a ROR record's external identifiers
}
\keyword{internal}
