% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_organisations.R
\name{ror_link_of_type}
\alias{ror_link_of_type}
\title{One of a ROR record's links}
\usage{
ror_link_of_type(record, type)
}
\arguments{
\item{record}{A ROR record, or `NULL`}

\item{type}{The link type, e.g. "website" or "wikipedia"}
}
\value{
The link, or `NA_character_`
}
\description{
One of a ROR record's links
}
\keyword{internal}
