% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_create_filtered_register_csvs.R
\name{save_filtered_csv}
\alias{save_filtered_csv}
\title{Save Filtered DataFrame to CSV}
\usage{
save_filtered_csv(
  filtered_register,
  filter,
  table_name,
  filter_subcategory = NULL
)
}
\arguments{
\item{filtered_register}{DataFrame to be saved.}

\item{filter}{Type of filter applied (e.g., "codecheckers", "venues").}

\item{table_name}{Name of the table or filtered item.}

\item{filter_subcategory}{Optional subcategory for the output directory structure (e.g., venue type). Default is NULL.}
}
\description{
Saves a filtered DataFrame to a CSV file in an output directory determined by the filter type, 
table name, and an optional subcategory. Creates the directory if it doesn't exist.
}
