% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_register_json.R
\name{set_paper_title_references}
\alias{set_paper_title_references}
\title{Set "Title" and "Paper reference" columns and values to the register_table}
\usage{
set_paper_title_references(register_table)
}
\arguments{
\item{register_table}{The register table}
}
\value{
Updated register table including "Title" and "Paper reference" columns
}
\description{
Set "Title" and "Paper reference" columns and values to the register_table
}
