% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecheck.R
\name{set_zenodo_certificate}
\alias{set_zenodo_certificate}
\title{Upload the CODECHECK certificate to Zenodo.}
\usage{
set_zenodo_certificate(zen, record, certificate)
}
\arguments{
\item{zen}{- Object from zen4R to interact with Zenodo}

\item{certificate}{name of the PDF file.}

\item{report}{- string containing the report URL on Zenodo.}
}
\description{
Upload the CODECHECK certificate to Zenodo.
}
\details{
Upload the CODECHECK certificate to Zenodo as a draft.  Warning: if
the file has already been uploaded once, you will need to delete it via
the web interface before being able to upload a new versin.
}
\author{
Stephen Eglen
}
