% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_build_metadata.R
\name{truncate_text}
\alias{truncate_text}
\title{Shorten text to a maximum length, at a word boundary}
\usage{
truncate_text(x, max_chars)
}
\arguments{
\item{x}{The text}

\item{max_chars}{Maximum number of characters}
}
\value{
The text, shortened and suffixed with a horizontal ellipsis if needed
}
\description{
Shorten text to a maximum length, at a word boundary
}
\keyword{internal}
