% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_build_metadata.R
\name{venue_schema_org_type}
\alias{venue_schema_org_type}
\title{Map a CODECHECK venue type to the closest Schema.org entity type}
\usage{
venue_schema_org_type(venue_type)
}
\arguments{
\item{venue_type}{journal/conference/community/institution (register.csv `Type`)}
}
\value{
A Schema.org `@type` string
}
\description{
Map a CODECHECK venue type to the closest Schema.org entity type
}
\keyword{internal}
