% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikibase_queries.R
\name{wikibase_local_ids}
\alias{wikibase_local_ids}
\title{The instance's local id for each Wikidata id}
\usage{
wikibase_local_ids(plan)
}
\arguments{
\item{plan}{a plan from [plan_wikibase_entities()] with `local_id` filled in}
}
\value{
a named character vector, Wikidata id to local id
}
\description{
The instance's local id for each Wikidata id
}
\keyword{internal}
