% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikibase_pages.R
\name{wikibase_page_url}
\alias{wikibase_page_url}
\title{A page's address on the instance}
\usage{
wikibase_page_url(title)
}
\arguments{
\item{title}{the page title}
}
\value{
the URL
}
\description{
A page's address on the instance
}
\keyword{internal}
