% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikibase_export.R
\name{wikibase_snak}
\alias{wikibase_snak}
\title{Build one snak}
\usage{
wikibase_snak(property, datatype, value)
}
\arguments{
\item{property}{the *local* property id on the instance}

\item{datatype}{the property's datatype}

\item{value}{the value}
}
\description{
Build one snak
}
\keyword{internal}
