% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikibase.R
\name{wikibase_sparql}
\alias{wikibase_sparql}
\title{Run a SPARQL query against Wikidata or the CODECHECK Wikibase}
\usage{
wikibase_sparql(query, endpoint = c("wikibase", "main", "scholarly"))
}
\arguments{
\item{query}{the query text}

\item{endpoint}{one of [WIKIDATA_ENDPOINTS], by name}
}
\value{
a `data.frame` of the bindings, one column per selected variable
}
\description{
Run a SPARQL query against Wikidata or the CODECHECK Wikibase
}
\keyword{internal}
