% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikidata.R
\name{wikidata_creates}
\alias{wikidata_creates}
\title{Whether the export may create this kind of entity on a target}
\usage{
wikidata_creates(kind, target = c("wikidata", "wikibase"))
}
\arguments{
\item{kind}{one of [wikidata_entity_kinds()]}

\item{target}{`"wikidata"` or `"wikibase"`}
}
\value{
`TRUE` if the export may create such an item on that target
}
\description{
Only certificates are created on Wikidata: the papers, people and venues they
refer to are resolved there and created by the communities that own them. Our
own Wikibase mirrors everything, since it has no notability rules to respect.
}
\examples{
wikidata_creates("certificate", "wikidata")
wikidata_creates("paper", "wikidata")
}
