% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikidata.R
\name{wikidata_entity_kinds}
\alias{wikidata_entity_kinds}
\title{The entity kinds of the model}
\usage{
wikidata_entity_kinds()
}
\value{
a character vector, e.g. `c("certificate", "paper", ...)`
}
\description{
The entity kinds of the model
}
