% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikidata.R
\name{wikidata_statements}
\alias{wikidata_statements}
\title{The statement definitions of one entity kind}
\usage{
wikidata_statements(kind)
}
\arguments{
\item{kind}{one of [wikidata_entity_kinds()]}
}
\value{
a list of statement definitions
}
\description{
The statement definitions of one entity kind
}
