% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikibase_queries.R
\name{write_wikibase_examples_page}
\alias{write_wikibase_examples_page}
\title{Write the example-queries page onto the instance}
\usage{
write_wikibase_examples_page(session, plan)
}
\arguments{
\item{session}{a session from [wikibase_session()]}

\item{plan}{a plan with `local_id` filled in}
}
\value{
the page title, invisibly
}
\description{
Write the example-queries page onto the instance
}
\keyword{internal}
