% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_cert_md.R
\name{add_abstract}
\alias{add_abstract}
\title{Inserts the abstract text and source link into the Markdown content if an abstract is found for the given repository. 
If no abstract is found, an empty string is inserted in place of the abstract content.}
\usage{
add_abstract(repo_link, md_content)
}
\arguments{
\item{repo_link}{A character string containing the repository link from which to retrieve the abstract.}

\item{md_content}{A character string containing the Markdown content with placeholders for abstract details.}
}
\value{
The markdown content with filled abstract placeholder
}
\description{
Inserts the abstract text and source link into the Markdown content if an abstract is found for the given repository. 
If no abstract is found, an empty string is inserted in place of the abstract content.
}
