% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_preprocess_register.R
\name{add_cert_links}
\alias{add_cert_links}
\title{Function for adding clickable links to the paper for each entry in the register table.}
\usage{
add_cert_links(register_table)
}
\arguments{
\item{register_table}{The register table to be adjusted.}
}
\value{
The adjusted register table with clickable Certificate links.
}
\description{
Function for adding clickable links to the paper for each entry in the register table.
}
