% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_cert_md.R
\name{add_codecheck_details_md}
\alias{add_codecheck_details_md}
\title{Populates an existing markdown content template with details about the codecheck details.}
\usage{
add_codecheck_details_md(md_content, repo_link)
}
\arguments{
\item{md_content}{A character string containing the Markdown template content with placeholders.}

\item{repo_link}{A character string containing the repository link associated with the certificate.}
}
\value{
The markdown content, with codecheck details placeholders filled.
}
\description{
Populates an existing markdown content template with details about the codecheck details.
}
