% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_register_mds.R
\name{add_markdown_title}
\alias{add_markdown_title}
\title{Function to add the markdown title based on the specific register table name.}
\usage{
add_markdown_title(table_details, md_table, filter)
}
\arguments{
\item{table_details}{List containing details such as the table name, subcat name.}

\item{md_table}{The markdown table where the title needs to be added}

\item{filter}{The filter}
}
\value{
The modified markdown table
}
\description{
Function to add the markdown title based on the specific register table name.
}
