% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_preprocess_register.R
\name{add_openalex_ids}
\alias{add_openalex_ids}
\title{Add OpenAlex work IDs to the register table (addresses register#185)}
\usage{
add_openalex_ids(register_table, register)
}
\arguments{
\item{register_table}{The register table}

\item{register}{The register from the register.csv file}
}
\value{
The register table with added "OpenAlex" column
}
\description{
Add OpenAlex work IDs to the register table (addresses register#185)
}
