% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_cert_md.R
\name{add_paper_details_md}
\alias{add_paper_details_md}
\title{Populates an existing markdown content template with details about the codechecked paper.}
\usage{
add_paper_details_md(md_content, repo_link, download_cert_status)
}
\arguments{
\item{md_content}{A character string containing the Markdown template content with placeholders.}

\item{repo_link}{A character string containing the repository link associated with the certificate.}
}
\value{
The markdown content, with paper details placeholders filled.
}
\description{
Populates an existing markdown content template with details about the codechecked paper.
}
