% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_preprocess_register.R
\name{add_report_links}
\alias{add_report_links}
\title{Function for adding report URLs for each entry in the register table.}
\usage{
add_report_links(register_table, register)
}
\arguments{
\item{register_table}{The register table}

\item{register}{The register from the register.csv file}
}
\value{
The adjusted register table with Report column containing plain URLs
}
\description{
Function for adding report URLs for each entry in the register table.
}
