% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{as_latex_url}
\alias{as_latex_url}
\title{Wrap URL for LaTeX}
\usage{
as_latex_url(x)
}
\arguments{
\item{x}{- A string that may contain URLs that should be hyperlinked.}
}
\value{
A string with the passed URL as a latex `\url{http://the.url}`
}
\description{
Wrap URL for LaTeX
}
\author{
Stephen Eglen
}
