% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{codecheck_metadata}
\alias{codecheck_metadata}
\title{Return the metadata for the codecheck project in root folder of project}
\usage{
codecheck_metadata(root = getwd())
}
\arguments{
\item{root}{Path to the root folder of the project, defaults to current working directory}
}
\value{
A list containing the metadata found in the codecheck.yml file
}
\description{
Return the metadata for the codecheck project in root folder of project
}
\details{
Loads and parses the codecheck.yml file from the specified directory.
If the file doesn't exist, stops with a clear error message.
}
\author{
Stephen Eglen
}
