% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_download_certs.R,
%   R/utils_render_cert_htmls.R
\name{convert_cert_pdf_to_png}
\alias{convert_cert_pdf_to_png}
\title{Converts each page of a certificate PDF to JPEG format images, saving them in the specified certificate directory.}
\usage{
convert_cert_pdf_to_png(cert_id)

convert_cert_pdf_to_png(cert_id)
}
\arguments{
\item{cert_id}{The certificate identifier. This ID is used to locate the PDF and save the resulting images.}
}
\description{
Converts each page of a certificate PDF to JPEG format images, saving them in the specified certificate directory.

Converts each page of a certificate PDF to JPEG format images, saving them in the specified certificate directory.
}
