% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecheck.R
\name{copy_manifest_files}
\alias{copy_manifest_files}
\title{Copy files from manifest into the codecheck folder and summarise.}
\usage{
copy_manifest_files(root, metadata, dest_dir, keep_full_path = FALSE)
}
\arguments{
\item{root}{- Path to the root folder of the project.}

\item{metadata}{- the codecheck metadata list.}

\item{dest_dir}{- folder where outputs are to be copied to (codecheck/outputs)}

\item{keep_full_path}{- TRUE to keep relative pathname of figures.}
}
\value{
A dataframe containing one row per manifest file.
}
\description{
Copy manifest files into the root/codecheck/outputs folder; return manifest.
}
\details{
The metadata should specify the manifest -- the files to copy into the
codecheck/oututs folder.  Each of the files in the manifest is copied into
the destination directory and then the manifest is returned as a dataframe.
If KEEP_FULL_PATH is TRUE, we keep the full path for the output files.
This is useful when there are two output files with the same name in
different folders, e.g. expt1/out.pdf and expt2/out.pdf
}
\author{
Stephen Eglen
}
