% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_external_libs.R
\name{copy_register_css}
\alias{copy_register_css}
\title{Copy the CODECHECK Register CSS to the Output Assets Directory}
\usage{
copy_register_css(assets_dir)
}
\arguments{
\item{assets_dir}{Directory for the register's own assets, by convention a
sibling of the libraries directory, see [register_assets_dir()]}
}
\description{
Copy the CODECHECK Register CSS to the Output Assets Directory
}
\keyword{internal}
