% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_cert_md.R
\name{create_cert_md}
\alias{create_cert_md}
\title{Generates a Markdown file for a certificate based on a specified template, filling in details about the 
paper, authors, codecheck information, and the certificate images if available. The resulting Markdown file is later rendered to HTML.}
\usage{
create_cert_md(cert_id, repo_link, download_cert_status)
}
\arguments{
\item{cert_id}{A character string representing the unique identifier of the certificate.}

\item{repo_link}{A character string containing the repository link associated with the certificate.}

\item{download_cert_status}{An integer (0 or 1) indicating whether the certificate PDF was downloaded (1) or not (0).}
}
\description{
Generates a Markdown file for a certificate based on a specified template, filling in details about the 
paper, authors, codecheck information, and the certificate images if available. The resulting Markdown file is later rendered to HTML.
}
