% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_cert_htmls.R
\name{create_cert_page_section_files}
\alias{create_cert_page_section_files}
\title{Generates section files for a certificate HTML page, including prefix, postfix, and header HTML components.}
\usage{
create_cert_page_section_files(output_dir)
}
\arguments{
\item{output_dir}{A character string specifying the directory where the section files will be saved.}
}
\description{
Generates section files for a certificate HTML page, including prefix, postfix, and header HTML components.
}
