% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{create_codecheck_files}
\alias{create_codecheck_files}
\title{Create template files for the codecheck process.}
\usage{
create_codecheck_files()
}
\value{
Nothing
}
\description{
Create template files for the codecheck process.
}
\details{
This function simply creates some template files to help start the
codecheck process.  If either ./codecheck.yml or codecheck/ exists then
it assumes you have already started codechecking, and so will not copy any
files across.
}
\author{
Stephen J. Eglen
}
