% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_reigster_html.r
\name{create_index_postfix_html}
\alias{create_index_postfix_html}
\title{Dynamically generates the index_postfix.html from a template file}
\usage{
create_index_postfix_html(
  output_dir,
  filter,
  register_table_name,
  is_reg_table
)
}
\arguments{
\item{output_dir}{The output directory}

\item{filter}{The filter name}

\item{register_table_name}{The register table name. If this is NULL we are generating list of venues/ codecheckers}
}
\description{
Dynamically generates the index_postfix.html from a template file
}
