% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_reigster_html.r
\name{create_index_postfix_html}
\alias{create_index_postfix_html}
\title{Dynamically generates the index_postfix.html from a template file}
\usage{
create_index_postfix_html(filter, register_table_name)
}
\arguments{
\item{filter}{The filter name}

\item{register_table_name}{The register table name}
}
\description{
Dynamically generates the index_postfix.html from a template file
}
