% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_register_htmls.R
\name{create_index_prefix_html}
\alias{create_index_prefix_html}
\title{Dynamically generates the index_prefix.html from a template file}
\usage{
create_index_prefix_html(output_dir)
}
\arguments{
\item{output_dir}{The output directory}
}
\description{
Dynamically generates the index_prefix.html from a template file
}
