% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_reigster_html.r
\name{create_index_section_files}
\alias{create_index_section_files}
\title{Creates index postfix, prefix and the header}
\usage{
create_index_section_files(
  output_dir,
  filter,
  register_table_name,
  is_reg_table
)
}
\arguments{
\item{output_dir}{The output directory of the section files}

\item{filter}{The filter name}

\item{register_table_name}{The register table name}
}
\description{
Creates index postfix, prefix and the header
}
