% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_register_htmls.R
\name{create_index_section_files}
\alias{create_index_section_files}
\title{Creates index postfix, prefix and the header}
\usage{
create_index_section_files(output_dir, filter, table_details)
}
\arguments{
\item{output_dir}{The output directory of the section files}

\item{filter}{The filter name}

\item{table_details}{List containing details such as the table name, subcat name.}
}
\description{
Creates index postfix, prefix and the header
}
