% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_register_htmls.R
\name{create_index_section_files}
\alias{create_index_section_files}
\title{Creates index postfix, prefix and the header}
\usage{
create_index_section_files(
  output_dir,
  filter,
  register_table_name,
  is_reg_table,
  filter_subcategory = NULL
)
}
\arguments{
\item{output_dir}{The output directory of the section files}

\item{filter}{The filter name}

\item{register_table_name}{The register table name}

\item{filter_subcategory}{The name of the filter subcategory. Only needed in case of venues which has subcategories. Defaults to NULL}
}
\description{
Creates index postfix, prefix and the header
}
