% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_reigster_html.r
\name{create_index_section_files}
\alias{create_index_section_files}
\title{Creates html files for each index section- the index postfix, prefix and the header}
\usage{
create_index_section_files(filter, register_table_name)
}
\arguments{
\item{filter}{The filter name}

\item{register_table_name}{The register table name}
}
\description{
Creates html files for each index section- the index postfix, prefix and the header
}
