% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_register_mds.R
\name{create_md_table}
\alias{create_md_table}
\title{Creates a markdown table from a register template
Adds title to the markdown and adjusts the column widths of the table 
before returning it.}
\usage{
create_md_table(register_table, table_details, filter)
}
\arguments{
\item{register_table}{DataFrame of the register data.}

\item{table_details}{List containing details such as the table name, subcat name.}

\item{filter}{Type of filter (e.g., "venues", "codecheckers").}
}
\value{
The markdown table
}
\description{
Creates a markdown table from a register template
Adds title to the markdown and adjusts the column widths of the table 
before returning it.
}
