% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_download_certs.R
\name{extract_cert_pdf_from_zip}
\alias{extract_cert_pdf_from_zip}
\title{Downloads a ZIP file from the given URL, searches for "codecheck.pdf" within its contents,
renames it to "cert.pdf," and saves it in the specified directory.}
\usage{
extract_cert_pdf_from_zip(zip_download_url, cert_sub_dir)
}
\arguments{
\item{zip_download_url}{URL to download the ZIP file from.}

\item{cert_sub_dir}{Directory to save the extracted certificate PDF.}
}
\value{
1 if "codecheck.pdf" is found and saved, otherwise 0.
}
\description{
Downloads a ZIP file from the given URL, searches for "codecheck.pdf" within its contents,
renames it to "cert.pdf," and saves it in the specified directory.
}
