% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_cert_htmls.R
\name{generate_cert_json}
\alias{generate_cert_json}
\title{Generates a JSON file with all certificate metadata}
\usage{
generate_cert_json(cert_id, repo_link, cert_type, cert_venue)
}
\arguments{
\item{cert_id}{A character string representing the unique identifier of the certificate.}

\item{repo_link}{A character string containing the repository link associated with the certificate.}

\item{cert_type}{A character string containing the venue type (journal, conference, community, institution).}

\item{cert_venue}{A character string containing the venue name.}
}
\description{
Creates an index.json file containing all information displayed on the
certificate landing page for programmatic access.
}
