% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_reigster_html.r
\name{generate_html_document_yml}
\alias{generate_html_document_yml}
\title{Dynamically generates a html_document.yml with the full paths to the index header, prefix 
and postfix.html files.}
\usage{
generate_html_document_yml(filter, register_table_name)
}
\arguments{
\item{filter}{The filter name}

\item{register_table_name}{The register table name}
}
\description{
Dynamically generates a html_document.yml with the full paths to the index header, prefix 
and postfix.html files.
}
