% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_register_htmls.R
\name{generate_html_postfix_hrefs_reg}
\alias{generate_html_postfix_hrefs_reg}
\title{Generates the hrefs to set in the postfix.html file for the rendering of register tables.}
\usage{
generate_html_postfix_hrefs_reg(filter, table_details)
}
\arguments{
\item{filter}{The filter name}

\item{table_details}{List containing details such as the table name, subcat name.}
}
\value{
A list with the hrefs
}
\description{
Generates the hrefs to set in the postfix.html file for the rendering of register tables.
}
