% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_reigster_html.r
\name{generate_html_postfix_hrefs_reg}
\alias{generate_html_postfix_hrefs_reg}
\title{Generates the hrefs to set in the postfix.html file for the rendering of register tables.}
\usage{
generate_html_postfix_hrefs_reg(filter, register_table_name)
}
\arguments{
\item{filter}{The filter name}

\item{register_table_name}{The register table name}
}
\description{
Generates the hrefs to set in the postfix.html file for the rendering of register tables.
}
