% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_table_non_registers.R
\name{generate_html_title_non_registers}
\alias{generate_html_title_non_registers}
\title{Generate HTML Title for Non-Register Pages}
\usage{
generate_html_title_non_registers(filter, subcat)
}
\arguments{
\item{filter}{A string specifying the filter applied (e.g., "venues", "codecheckers").}

\item{subcat}{An optional string for the subcategory (if applicable).}
}
\value{
A string containing the generated HTML title.
}
\description{
Generates the title for the HTML page based on the filter and subcategory.
The title is used for rendering the non-register HTML files such as venues or codecheckers pages.
}
