% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_codecheckers.R
\name{get_institutional_codecheckers_data}
\alias{get_institutional_codecheckers_data}
\title{Fetch and cache institutional-codecheckers.csv data from GitHub}
\usage{
get_institutional_codecheckers_data()
}
\value{
A data frame with columns: name, handle, ORCID, contact, fields, languages
  (the last three are `NA` - the institution is the point of contact)
}
\description{
Fetch and cache institutional-codecheckers.csv data from GitHub
}
