% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ror.R
\name{get_openalex_publication_date_cached}
\alias{get_openalex_publication_date_cached}
\title{Cached version of [get_openalex_publication_date_result()]}
\usage{
get_openalex_publication_date_cached(openalex_id)
}
\arguments{
\item{openalex_id}{An OpenAlex work URL or ID, e.g.
"https://openalex.org/W3014157798"}
}
\value{
The publication date as a character string, or `NA_character_`
}
\description{
Cached version of [get_openalex_publication_date_result()]
}
\keyword{internal}
