% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_create_filtered_register_csvs.R
\name{get_output_dir}
\alias{get_output_dir}
\title{Gets the output dir depending on the filter name and the value of the filtered column}
\usage{
get_output_dir(filter, column_value)
}
\arguments{
\item{filter}{The filter name}

\item{column_value}{The value of the column the filter applies to}
}
\value{
The directory to save files to
}
\description{
Gets the output dir depending on the filter name and the value of the filtered column
}
