% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/researchequals.R
\name{get_researchequals_version_metadata}
\alias{get_researchequals_version_metadata}
\title{Retrieve a ResearchEquals version's metadata}
\usage{
get_researchequals_version_metadata(version_id)
}
\arguments{
\item{version_id}{the ResearchEquals version ID}
}
\value{
the parsed API response as a list
}
\description{
Retrieve the metadata of a ResearchEquals module version
}
\author{
Daniel Nuest
}
