% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecheck.R
\name{get_zenodo_record}
\alias{get_zenodo_record}
\title{Get the full zenodo record using the record number stored in the metadata.}
\usage{
get_zenodo_record(zenodo, metadata)
}
\arguments{
\item{zenodo}{}

\item{metadata}{}
}
\value{
The Zenodo record, or NULL.
}
\description{
Get the full Zenodo record from the metadata
}
\details{
Retrieve the zenodo record, if one exists.
If no record number is currently  listed in the metadata (i.e. the "FIXME" tag is still there)
then the code returns NULL and an empty record should be created.
}
\author{
Stephen Eglen
}
