% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{latex_summary_of_manifest}
\alias{latex_summary_of_manifest}
\title{Print a latex table to summarise CODECHECK metadata}
\usage{
latex_summary_of_manifest(
  metadata,
  manifest_df,
  root,
  align = c("l", "p{6cm}", "p{6cm}", "p{2cm}")
)
}
\arguments{
\item{metadata}{- the CODECHECK metadata list.}

\item{manifest_df}{- The manifest data frame}

\item{root}{- root directory of the project}

\item{align}{- alignment flags for the table.}
}
\value{
The latex table, suitable for including in the Rmd
}
\description{
Print a latex table to summarise CODECHECK manfiest
}
\details{
Format a latex table that summarises the main CODECHECK manifest
}
\author{
Stephen Eglen
}
