% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecheck.R
\name{latex_summary_of_manifest}
\alias{latex_summary_of_manifest}
\title{Print a latex table to summarise CODECHECK metadata}
\usage{
latex_summary_of_manifest(manifest_df)
}
\arguments{
\item{manifest_df}{- The manifest data frame}
}
\value{
The latex table, suitable for including in the Rmd
}
\description{
Print a latex table to summarise CODECHECK manfiest
}
\details{
Format a latex table that summarises the main CODECHECK manifest
}
\author{
Stephen Eglen
}
