% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{register_check}
\alias{register_check}
\title{Function for checking all entries in the register}
\usage{
register_check(register = read.csv("register.csv", as.is = TRUE))
}
\arguments{
\item{register}{A `data.frame` with all required information for the register's view}
}
\description{
This functions starts of a `data.frame` read from the local register file.
}
\details{
Futher test ideas:

- Does the repo have a LICENSE?
}
