% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{register_clear_cache}
\alias{register_clear_cache}
\title{Function for clearing the register cache}
\usage{
register_clear_cache()
}
\value{
0 for success, 1 for failure, invisibly (see `unlink`)
}
\description{
Function for clearing the register cache
}
\author{
Daniel Nüst
}
