% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_cert_htmls.R
\name{render_cert_htmls}
\alias{render_cert_htmls}
\title{Generates HTML files for each certificate listed in the given register table. 
It checks for the existence of the certificate PDF, downloads it if necessary, and 
converts it to JPEG format for embedding.}
\usage{
render_cert_htmls(register_table, force_download = FALSE)
}
\arguments{
\item{register_table}{A data frame containing details of each certificate, including repository links and report links.}

\item{force_download}{Logical; if TRUE, forces the download of certificate PDFs even if they already exist locally. Defaults to FALSE.}
}
\description{
Generates HTML files for each certificate listed in the given register table. 
It checks for the existence of the certificate PDF, downloads it if necessary, and 
converts it to JPEG format for embedding.
}
