% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_register_htmls.R
\name{render_html}
\alias{render_html}
\title{Renders html for a single table}
\usage{
render_html(table, table_details, filter)
}
\arguments{
\item{table}{The table to render to HTML}

\item{table_details}{List containing details such as the table name, subcat name.}

\item{filter}{The filter}
}
\description{
This function can be used to render both register and non-register tables
}
