% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_register_general.r
\name{render_register}
\alias{render_register}
\title{Render Register in Specified Output Format}
\usage{
render_register(register_table, table_details, filter, output_type)
}
\arguments{
\item{register_table}{The register table that needs to be rendered into different files.}

\item{table_details}{A list of details related to the table (e.g., output directory, metadata).}

\item{filter}{A string specifying the filter applied to the register data.}

\item{output_type}{A string specifying the desired output format ("md" for Markdown, "html" for HTML, "json" for JSON).}
}
\value{
None. The function generates a file in the specified format.
}
\description{
This function renders the register table into different output formats based on the specified type.
It supports rendering the table as Markdown, HTML, or JSON.
}
