% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_register_htmls.R
\name{render_register_html}
\alias{render_register_html}
\title{Renders register html for a single register_table}
\usage{
render_register_html(
  filter,
  register_table,
  register_table_name,
  filter_subcategory = NULL
)
}
\arguments{
\item{filter}{The filter}

\item{register_table}{The register table}

\item{register_table_name}{The register table name}

\item{filter_subcategory}{An optional string representing a subcategory within the filter 
(e.g., venue type). Default is NULL.}
}
\description{
Renders register html for a single register_table
}
