% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_register.R
\name{render_register_html}
\alias{render_register_html}
\title{Function for rendering the register html file. 
The html file is rendered from the markdown file.}
\usage{
render_register_html(list_register_tables, md_columns_widths)
}
\arguments{
\item{md_columns_widths}{The column widths for the markdown file}

\item{register_table}{The register table}
}
\value{
None
}
\description{
Function for rendering the register html file. 
The html file is rendered from the markdown file.
}
