% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_register.R
\name{render_register_json}
\alias{render_register_json}
\title{Function for rendering the register json file.}
\usage{
render_register_json(list_register_tables, register)
}
\arguments{
\item{register}{The register from the register.csv file}

\item{register_table}{The register table}
}
\value{
None
}
\description{
Function for rendering the register json file.
}
