% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_register_json.R
\name{render_register_json}
\alias{render_register_json}
\title{Renders register json for a single register_table}
\usage{
render_register_json(register_table, table_details, filter)
}
\arguments{
\item{register_table}{The register table}

\item{table_details}{List containing details such as the table name, subcat name.}

\item{filter}{The filter}
}
\description{
Renders register json for a single register_table
}
