% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_register_mds.R
\name{render_register_md}
\alias{render_register_md}
\title{Renders register md for a single register_table}
\usage{
render_register_md(
  register_table,
  table_details,
  filter,
  for_html_file = FALSE
)
}
\arguments{
\item{register_table}{The register table}

\item{table_details}{List containing details such as the table name, subcat name.}

\item{filter}{The filter}

\item{for_html_file}{Flag for whether we are rendering register md for html file.
Set to FALSE by default. If TRUE, no repo links are added to the repository table.}
}
\description{
Renders register md for a single register_table
}
