% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_register_mds.R
\name{render_register_md}
\alias{render_register_md}
\title{Renders register md for a single register_table}
\usage{
render_register_md(
  filter,
  register_table,
  register_table_name,
  filter_subcategory = NULL,
  for_html_file = FALSE
)
}
\arguments{
\item{filter}{The filter}

\item{register_table}{The register table}

\item{register_table_name}{The register table name}

\item{filter_subcategory}{The name of the filter subcategory. Only needed in case of venues which has subcategories. Defaults to NULL}

\item{for_html_file}{Flag for whether we are rendering register md for html file.
Set to FALSE by default. If TRUE, no repo links are added to the repository table.}
}
\description{
Renders register md for a single register_table
}
