% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_register_mds.R
\name{render_register_md}
\alias{render_register_md}
\title{Renders register md for a single register_table}
\usage{
render_register_md(
  filter,
  register_table,
  register_table_name,
  is_temp_file = FALSE
)
}
\arguments{
\item{filter}{The filter}

\item{register_table}{The register table}

\item{register_table_name}{The register table name}
}
\description{
Renders register md for a single register_table
}
