% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_render_register_mds.R
\name{save_md_table}
\alias{save_md_table}
\title{Save markdown table to a file}
\usage{
save_md_table(output_dir, md_table, for_html_file)
}
\arguments{
\item{output_dir}{The output_dir}

\item{md_table}{The markdown table to be saved.}

\item{for_html_file}{A logical flag indicating whether the markdown is being rendered for an HTML file. 
If TRUE, the file is saved as `temp.md`. Default is FALSE.}
}
\description{
The file is saved as either a temporary file (`temp.md`) or as `register.md` depending on 
whether it is being rendered for an HTML file.
}
