% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codecheck.R
\name{upload_zenodo_metadata}
\alias{upload_zenodo_metadata}
\title{Upload metadata to Zenodod}
\usage{
upload_zenodo_metadata(zenodo, myrec, metadata)
}
\arguments{
\item{zenodo}{object from zen4R to connect with Zenodo}

\item{myrec}{a Zenodo record object}

\item{metadata}{codecheck metadata, likely loaded from a codecheck.yml file}
}
\value{
rec -- the updated record.
}
\description{
Upload codecheck metadata to Zenodo.
}
\details{
The contents of codecheck.yml are uploaded to Zenodo using this funciton.
}
\author{
Stephen Eglen
}
