% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{validate_codecheck_yml}
\alias{validate_codecheck_yml}
\title{Validate a CODECHECK configuration}
\usage{
validate_codecheck_yml(configuration)
}
\arguments{
\item{configuration}{R object of class `yaml`}
}
\value{
`TRUE` if the provided configuration is valid, `FALSE` otherwise.
}
\description{
This functions checks "MUST"-contents only, see https://codecheck.org.uk/spec/config/latest/
}
\details{
For each entry in the registry we can do some basic checks, e.g. like
- checking the certificate numbers match
- checking the yaml exists?
- linting the yaml?
}
