% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{validate_codecheck_yml}
\alias{validate_codecheck_yml}
\title{Validate a CODECHECK configuration}
\usage{
validate_codecheck_yml(configuration)
}
\arguments{
\item{configuration}{R object of class `list`, or a path to a file}
}
\value{
`TRUE` if the provided configuration is valid, otherwise the function stops with an error
}
\description{
This functions checks "MUST"-contents only, see https://codecheck.org.uk/spec/config/latest/
}
\author{
Daniel Nuest
}
